banner module postman test complete

+++ Backend/src/app/module/banner/banner.service.ts
@@ -1,5 +1,8 @@
 import { Banner } from "./banner.model";
 import { IBanner } from "./banner.interface";
+import AppError from "../../errorHelper/AppError";
+import httpStatus from "http-status-codes";
+import { deleteImageFromCLoudinary } from "../../config/cloudinary.config";
 
 const createBanner = async (payload: IBanner) => {
   const banner = await Banner.create(payload);
@@ -26,10 +29,26 @@ const getBannerById = async (id: string) => {
 };
 
 const updateBanner = async (id: string, payload: Partial<IBanner>) => {
-  return await Banner.findByIdAndUpdate(id, payload, { new: true });
+  const existingBanner = await Banner.findById(id);
+  if (!existingBanner) {
+    throw new AppError(httpStatus.NOT_FOUND, "Banner not found");
+  }
+
+  const updatedBanner = await Banner.findByIdAndUpdate(id, payload, {
+    new: true,
+  });
+  if (payload.image && existingBanner.image) {
+    await deleteImageFromCLoudinary(existingBanner.image);
+  }
+  return updatedBanner;
 };
 
 const deleteBanner = async (id: string) => {
+  const banner = await Banner.findById(id);
+  if (!banner) {
+    throw new AppError(httpStatus.NOT_FOUND, "Banner not found");
+  }
+  await deleteImageFromCLoudinary(banner.image);
   return await Banner.findByIdAndDelete(id);
 };
